+++ examples/use_memmap.py
@@ -65,7 +65,7 @@ if __name__ == '__main__':
         # reconstruct, write res out to a memmap with name res_file
         grappa(
             kspace, calib, kernel_size, coil_axis=-1, lamda=0.01,
-            disp=False, memmap=True, memmap_filename=res_file)
+            memmap=True, memmap_filename=res_file)
 
         # Take a look by opening up the memmap
         res = np.memmap(
@@ -73,11 +73,15 @@ if __name__ == '__main__':
             dtype='complex')
         res = np.abs(np.sqrt(N**2)*np.fft.fftshift(np.fft.ifft2(
             np.fft.ifftshift(res, axes=ax), axes=ax), axes=ax))
-        res0 = np.zeros((2*N, 2*N))
-        kk = 0
-        for idx in np.ndindex((2, 2)):
-            ii, jj = idx[:]
-            res0[ii*N:(ii+1)*N, jj*N:(jj+1)*N] = res[..., kk]
-            kk += 1
-        plt.imshow(res0, cmap='gray')
-        plt.show()
+
+        from mr_utils import view
+        view(res)
+
+        # res0 = np.zeros((2*N, 2*N))
+        # kk = 0
+        # for idx in np.ndindex((2, 2)):
+        #     ii, jj = idx[:]
+        #     res0[ii*N:(ii+1)*N, jj*N:(jj+1)*N] = res[..., kk]
+        #     kk += 1
+        # plt.imshow(res0, cmap='gray')
+        # plt.show()
