+++ pygrappa/__init__.py
@@ -1,3 +1,4 @@
 '''Bring functions up to the correct level.'''
 
+from .lustig_grappa import lustig_grappa
 from .grappa import grappa
