+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import find_packages
 
 setup(
     name='pygrappa',
-    version='0.0.3',
+    version='0.1.0',
     author='Nicholas McKibben',
     author_email='nicholas.bgp@gmail.com',
     packages=find_packages(),
